Sending attachments via the SMTP client has been created.

+++ src/main/java/ru/dlabs/library/email/converter/MessagePartConverter.java
@@ -66,7 +66,7 @@ public class MessagePartConverter {
                 .data(content)
                 .type(AttachmentType.find(part.getContentType()))
                 .contentType(EmailMessageUtils.decodeData(part.getContentType()))
-                .size(content.length)
+                .size((long) content.length)
                 .build();
         } catch (MessagingException e) {
             throw new ReadMessageException(
